REFACTOR: Refatorando de forma a diminuir duplicidade de código entre os motores e passando de Abstract Factory para factory.
